Remove Cake intellisense, Splunk and ngrok
I haven't used Cake intellisense in years. The instructions are
outdated.

I haven't used Splunk in years, the instructions are most likely
outdated.

ngrok Basic tier (a paid product) is now required for it to be useful. I
also haven't used the product in years so I can't really recommend it.